load: split postgresql into different files
