+++ invenio_rdm_migrator/load/postgresql/__init__.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2023 CERN.
+#
+# Invenio-RDM-Migrator is free software; you can redistribute it and/or modify
+# it under the terms of the MIT License; see LICENSE file for more details.
+
+"""Invenio RDM migration PostgreSQL load module."""
+
+from .copy import PostgreSQLCopyLoad
+from .table_generators import (
+    ExistingDataTableGenerator,
+    SingleTableGenerator,
+    TableGenerator,
+)
+
+__all__ = (
+    "ExistingDataTableGenerator",
+    "PostgreSQLCopyLoad",
+    "SingleTableGenerator",
+    "TableGenerator",
+)
